+++ src/Portenta_H7_PWM_Impl.h
@@ -6,7 +6,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/Portenta_H7_PWM
   Licensed under MIT license
 
-  Version: 2.0.3
+  Version: 2.1.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -15,10 +15,13 @@
   2.0.1   K.Hoang      11/12/2021 Fix PWM_Multi example. Temporary fix polarity for HRTIM PWM
   2.0.2   K.Hoang      07/02/2022 Convert to h-only. Optimize code.
   2.0.3   K.Hoang      21/01/2023 Add `PWM_StepperControl` example
+  2.1.0   K.Hoang      24/01/2023 Add `PWM_manual` example and related function. Optimize for speed
 *****************************************************************************************************************************/
 
 #ifndef Portenta_H7_PWM_IMPL_H
 #define Portenta_H7_PWM_IMPL_H
+
+///////////////////////////////////////////
 
 #ifndef _PWM_LOGLEVEL_
   #define _PWM_LOGLEVEL_       1
@@ -29,6 +32,7 @@
 
 #include "PWM_Generic_Debug.h"
 
+///////////////////////////////////////////
 
 bool isValidPWMPin(const pin_size_t& pin)
 {
@@ -42,6 +46,8 @@ bool isValidPWMPin(const pin_size_t& pin)
   return false;
 }
 
+///////////////////////////////////////////
+
 bool isValidPWMDutyCycle(const pin_size_t& pin, const float& dutyCycle)
 {
   if ( (dutyCycle < 0.0f) || (dutyCycle > 100.0f) )
@@ -53,6 +59,8 @@ bool isValidPWMDutyCycle(const pin_size_t& pin, const float& dutyCycle)
 
   return true;
 }
+
+///////////////////////////////////////////
 
 bool isUsingHRTIM(const pin_size_t& pin)
 {
@@ -66,8 +74,12 @@ bool isUsingHRTIM(const pin_size_t& pin)
   return false;
 }
 
+///////////////////////////////////////////
+
 // Will calculate accurate min freq if necessary, depending on Portenta Clock
 #define MIN_HRTIM_PWM_FREQ      800
+
+///////////////////////////////////////////
 
 // Check if freq > 0 and if HRTIM => freq >= MIN_HRTIM_PWM_FREQ
 bool isValidPWMFreq(const pin_size_t& pin, const float& frequency)
@@ -88,6 +100,8 @@ bool isValidPWMFreq(const pin_size_t& pin, const float& frequency)
 
   return true;
 }
+
+///////////////////////////////////////////
 
 bool isValidPWMSettings(const pin_size_t& pin, const float& frequency, const float& dutyCycle)
 {
@@ -118,17 +132,16 @@ bool isValidPWMSettings(const pin_size_t& pin, const float& frequency, const flo
   return true;
 }
 
+///////////////////////////////////////////
+
 // dutyCycle from 0.0f to 100.0f
 mbed::PwmOut* setPWM(mbed::PwmOut* &pwm, const pin_size_t& pin, const float& frequency, const float& dutyCycle)
 {
-  PWM_LOGDEBUG7("Freq = ", frequency, ", \tDutyCycle = ", dutyCycle, ", \tDutyCycle % = ", dutyCycle / 100, ", \tPin = ",
+  PWM_LOGDEBUG7("Freq = ", frequency, ", \tDutyCycle % = ", dutyCycle, ", \tDutyCycle = ", dutyCycle / 100, ", \tPin = ",
                 pin);
 
-  //if ( !isValidPWMPin(pin) || !isValidPWMFreq(pin, frequency) || !isValidPWMDutyCycle(pin, dutyCycle) )
   if ( !isValidPWMSettings(pin, frequency, dutyCycle) )
   {
-    //PWM_LOGERROR("Bad pin, Freq or dutyCycle");
-
     return NULL;
   }
 
@@ -141,6 +154,7 @@ mbed::PwmOut* setPWM(mbed::PwmOut* &pwm, const pin_size_t& pin, const float& fre
     pwm = new mbed::PwmOut(digitalPinToPinName(pin));
 
     digitalPinToPwm(pin) = pwm;
+    
 
     pwm->period_us( 1000000.0f / frequency );
 
@@ -148,9 +162,13 @@ mbed::PwmOut* setPWM(mbed::PwmOut* &pwm, const pin_size_t& pin, const float& fre
   }
   else if (pwm && (digitalPinToPwm(pin) == pwm) )
   {
-    PWM_LOGDEBUG("Use existing pwm");
+    //PWM_LOGDEBUG("Use existing pwm");
 
-    pwm->period_us( 1000000.0f / frequency );
+    if ( frequency != 1000000.0f / pwm->read_period_us())
+    {
+      PWM_LOGDEBUG3("New Freq =", frequency, ", old freq =", 1000000.0f / pwm->read_period_us());
+      pwm->period_us( 1000000.0f / frequency );
+    }  
 
     pwm->write(percent);
   }
@@ -158,6 +176,21 @@ mbed::PwmOut* setPWM(mbed::PwmOut* &pwm, const pin_size_t& pin, const float& fre
   return pwm;
 }
 
+///////////////////////////////////////////
+
+// Must have same pin and frequency    
+// DCPercentage from 0.0f - 100.0f
+mbed::PwmOut* setPWM_DCPercentage_manual(mbed::PwmOut* &pwm, const pin_size_t& pin, const float& DCPercentage)
+{ 
+  PWM_LOGDEBUG5("DCPercentage % = ", DCPercentage, ", \tDutyCycle = ", DCPercentage / 100, ", \tPin = ", pin);
+                
+  pwm->write(DCPercentage / 100);
+  
+  return pwm;
+}
+
+///////////////////////////////////////////
+    
 mbed::PwmOut* stopPWM(mbed::PwmOut* &pwm, const pin_size_t& pin)
 {
   if (isUsingHRTIM(pin))